scripts/requirements.txt
elasticsearch==9.1.1
groq==0.32.0
matplotlib==3.10.7
minsearch==0.0.7
numpy==2.3.4
pandas==2.3.3
psutil==5.9.8
python-dotenv==1.1.1
qdrant_client==1.15.1
scikit_learn==1.7.2
seaborn==0.13.2
sentence_transformers==5.1.1
tqdm==4.67.1
